requirements.txt
aiogram==2.25.1
requests==2.28.2
python-dotenv==1.0.0
psycopg2-binary==2.9.5
